Fix race condition
When recreating a StatefulSet with a large volume claim, the status
is initialy uninitialized. This means the `replicas` field is 0
until the k8s controller manager runs and sets the status. If the
resize controller runs twice before the k8s controller manager is
able to update the status, we will initialize the backup before
the statefulset scaled down successfully.