Use storage class requested by statefulset instead of the old PVC
Until now we reused the storage class of the to be resized PVC. However
that can be an issue if the requested or default storage class changed
and the old storage class is not available anymore.

With this change we create new PVCs using the storage class (or the lack
of it, i.e. default) of the statefulset.